WebApp/requirements.txt
awsebcli==3.8.1
botocore==1.4.62
cement==2.8.2
click==6.6
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
Flask==0.11.1
Flask-Login==0.3.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
pathspec==0.3.4
python-dateutil==2.5.3
python-firebase==1.2
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.5
virtualenv==15.0.3
websocket-client==0.37.0
Werkzeug==0.11.11
